+++ TFShellExt/ThumbFishDocument.h
@@ -39,5 +39,5 @@ protected:
 	void SetSearchContent(CString& value);
 	virtual BOOL GetThumbnail(UINT, HBITMAP*, WTS_ALPHATYPE*);
 	virtual void OnDrawThumbnail(HDC hDrawDC, LPRECT lprcBounds);
-	int FindMolVersion(char* data, int dataLength, int lineNum);
+	int FindMolVersion(char* data, size_t dataLength, int lineNum);
 };
